refactoring models with shared data, reconfiguring account/profile to just profile pages
