+++ src/client/src/models/AccountMilestone.js
@@ -1,11 +1,11 @@
 export class AccountMilestone {
   constructor(data) {
     this.id = data.id
+    this.tier = data.tier
+    this.count = data.count
+    this.claimed = data.claimed
+    this.milestone = data.milestone
     this.accountId = data.accountId
     this.milestoneId = data.milestoneId
-    this.claimed = data.claimed
-    this.tier = data.tier
-    this.milestone = data.milestone
-    this.count = data.count
   }
 }
