+++ src/client/src/pages/ProfileBadges.vue
@@ -52,6 +52,7 @@
   </div>
 </template>
 
+
 <script>
 import { computed } from "vue";
 import { AppState } from '../AppState';
@@ -71,6 +72,7 @@ export default {
   components: { ChallengeBadge }
 }
 </script>
+
 
 <style scoped lang="scss">
 .badge-card {
