+++ src/client/src/pages/ProfileChallenges.vue
